+++ src/index.html
@@ -45,7 +45,6 @@
 
     <div class="container">
       <nav id="sidebar-nav">
-        <h2>Navigation</h2>
         <ul id="sidebar-list">
           <li class="loading">Loading navigation...</li>
         </ul>
